Add support for axial vectors and (pseudo)scalars
This is a major revision to the Feynman rule generation part of ChPTlib,
which is now written in a way that, rather having two hard-coded fields
(M and V), allows for the relatively simple addition of even more fields
in the future (tensor spurions? baryons?).
This breaks existing generated code, but the interface is unchanged
except that makevertex now takes many more arguments (one per field),
and that the conversion of the vector spurion to Q*A(mu) is only enabled
by the flag QED (i.e., QCD+QED mode).

Also: changes to names.hf to accommodate more fields, change to .mk
extension for Make include files