Add C statement for automatic main file edits
The intention is that once the main form file is working as intended,
the changes can be transferred to the .chpt source file so that the FORM
code can be regenerated and the same customizations be automatically
added to other diagram sets (e.g. other orders).
Also: modify verbose logging output, edit help messages